update business type done (#22)
